Rewrite manual page of typog-grep.
Split hard-wired or configured values
from those passed in via command-line options in typog-grep
to get more precise warnings (or errors).

Translate template file typog-grep.in with a
GNU autoconf alike mechanism to supply it with
the same version number as the LaTeX package.

Add an GNU Info generation path for typog-grep.pod
that can be used to check file/format integrity.